renamed courseSessionsStatus.js to enrollmentView.js, added enrollment balance to enrollment view #245, styled unenroll button
