+++ src/components/FeatureViews/Accounts/Accounts.js
@@ -270,6 +270,7 @@ class Accounts extends Component {
         }
 
         return (
+
             <Grid
                 className="Accounts"
                 item
