+++ src/components/FeatureViews/Registration/RegistrationCourse.js
@@ -27,7 +27,7 @@ import "./registration.scss";
 import {Link, Redirect, useRouteMatch} from "react-router-dom";
 import {stringToColor} from "components/FeatureViews/Accounts/accountUtils";
 import Loading from "../../Loading";
-import {DayConverter} from "../Accounts/TabComponents/CourseSessionStatus";
+import {DayConverter} from "../Accounts/TabComponents/EnrollmentView";
 import ConfirmIcon from "@material-ui/icons/CheckCircle";
 import UnconfirmIcon from "@material-ui/icons/Cancel"
 import {courseDateFormat} from "../../../utils";
