+++ src/theme/_colors.scss
@@ -36,4 +36,5 @@ $bioBackgroundColor: #F5F5F5;
 $goodStatusColor: #13D70F;
 $warningStatusColor: #F8D000;
 $badStatusColor: #FC1414;
+$badActionColor: #e74c3c;
 $neutralStatusColor: #34495e;
